adds in episode categories as a search term, enriches the data dictionary with lookup lists for categories and teams refs #1614
