+++ search/templates/search/data_dictionary_descriptions/episode/category.html
@@ -1,0 +1,10 @@
+The type of episode
+{% if data_dictionary.in_page %}
+  <a href=""
+     class="orange-link"
+     ng-click="open_modal('LookupListReferenceCtrl', '/templates/modals/lookuplist_reference.html', {lookuplist_name: 'Episode Category', lookuplist: extractSliceSchema.findField('episode', 'category_name').enum})">
+    category
+  </a>
+{% else %}
+  category, for example Inpatient, or OPAT
+{% endif %}
