+++ search/templates/search/widgets/many_to_many.html
