+++ search/tests/test_search_rule.py
@@ -36,6 +36,7 @@ class SearchRuleFieldTestCase(OpalTestCase):
             enum=[1, 2, 3],
             query_args=['value', 'query_type'],
             name="some_slug",
+            description="its a custom field",
             widget='some_widget.html',
             widget_description='partials/search/descriptions/widget_description.html',
             display_name='custom field you know',
@@ -173,194 +174,6 @@ class EpisodeSearchRuleTestCase(OpalTestCase):
             self.episode_rule.query(query_end)
 
 
-class EpisodeTeamQueryTestCase(OpalTestCase):
-    def setUp(self, *args, **kwargs):
-        super(EpisodeTeamQueryTestCase, self).setUp(*args, **kwargs)
-        _, self.episode_1 = self.new_patient_and_episode_please()
-        _, self.episode_2 = self.new_patient_and_episode_please()
-        _, self.episode_3 = self.new_patient_and_episode_please()
-        self.episode_rule = search_rules.EpisodeSearchRule(self.user)
-
-    def test_episode_team_wrong_query_param(self):
-        query_end = dict(
-            query_type="asdfsadf",
-            value=["Some Team"],
-            field="team"
-        )
-        with self.assertRaises(exceptions.SearchException) as er:
-            self.episode_rule.query(query_end)
-        self.assertEqual(
-            str(er.exception),
-            "unrecognised query type for the episode team query with asdfsadf"
-        )
-
-    def test_episode_team_all_of_one(self):
-        """
-            test all of with a single tag
-        """
-        self.episode_1.tagging_set.create(value="tree", archived=False)
-        self.episode_1.tagging_set.create(value="plant", archived=False)
-        self.episode_2.tagging_set.create(value="plant", archived=False)
-        self.episode_3.tagging_set.create(value="tree", archived=False)
-        query_end = dict(
-            query_type="All Of",
-            value=["Plant"],
-            field="team"
-        )
-        with patch.object(search_rule_fields.models.Tagging, "build_field_schema") as bfs:
-            bfs.return_value = [
-                dict(
-                    name="plant",
-                    title="Plant"
-                ),
-                dict(
-                    name="tree",
-                    title="Tree"
-                ),
-            ]
-            result = self.episode_rule.query(query_end)
-        self.assertEqual(len(result), 2)
-        self.assertEqual(result[0].id, self.episode_1.id)
-        self.assertEqual(result[1].id, self.episode_2.id)
-
-    def test_episode_team_all_of_archived(self):
-        """
-            test archived tags are returned
-        """
-        self.episode_1.tagging_set.create(value="plant", archived=True)
-        query_end = dict(
-            query_type="All Of",
-            value=["Plant"],
-            field="team"
-        )
-        with patch.object(search_rule_fields.models.Tagging, "build_field_schema") as bfs:
-            bfs.return_value = [
-                dict(
-                    name="plant",
-                    title="Plant"
-                ),
-                dict(
-                    name="tree",
-                    title="Tree"
-                ),
-            ]
-            result = self.episode_rule.query(query_end)
-        self.assertEqual(len(result), 1)
-        self.assertEqual(result[0].id, self.episode_1.id)
-
-    def test_episode_team_all_of_many(self):
-        """
-            test when looking for multiple tags only
-            episodes with all of those tags will be returned
-        """
-        self.episode_1.tagging_set.create(value="tree", archived=False)
-        self.episode_1.tagging_set.create(value="plant", archived=False)
-        self.episode_2.tagging_set.create(value="plant", archived=False)
-        self.episode_3.tagging_set.create(value="tree", archived=False)
-        query_end = dict(
-            query_type="All Of",
-            value=["Plant", "Tree"],
-            field="team"
-        )
-        with patch.object(
-            search_rule_fields.models.Tagging, "build_field_schema"
-        ) as bfs:
-            bfs.return_value = [
-                dict(
-                    name="plant",
-                    title="Plant"
-                ),
-                dict(
-                    name="tree",
-                    title="Tree"
-                ),
-            ]
-            result = self.episode_rule.query(query_end)
-        self.assertEqual(len(result), 1)
-        self.assertEqual(result[0].id, self.episode_1.id)
-
-    @patch("search.search_rule_fields.models.Tagging.build_field_schema")
-    def test_episode_team_enum(self, bfs):
-        bfs.return_value = [
-            dict(
-                name="plant",
-                title="Plant"
-            ),
-            dict(
-                name="tree",
-                title="Tree"
-            ),
-        ]
-        self.assertEqual(
-            ["Plant", "Tree"], self.episode_rule.get_field("team").enum
-        )
-
-    def test_episide_team_any_of_one(self):
-        """
-            test all of with a single tag
-        """
-        self.episode_1.tagging_set.create(value="tree", archived=False)
-        self.episode_1.tagging_set.create(value="plant", archived=False)
-        self.episode_2.tagging_set.create(value="plant", archived=False)
-        self.episode_3.tagging_set.create(value="tree", archived=False)
-        query_end = dict(
-            query_type="Any Of",
-            value=["Plant"],
-            field="team"
-        )
-        with patch.object(
-            search_rule_fields.models.Tagging, "build_field_schema"
-        ) as bfs:
-            bfs.return_value = [
-                dict(
-                    name="plant",
-                    title="Plant"
-                ),
-                dict(
-                    name="tree",
-                    title="Tree"
-                ),
-            ]
-            result = self.episode_rule.query(query_end)
-        self.assertEqual(len(result), 2)
-        self.assertEqual(result[0].id, self.episode_1.id)
-        self.assertEqual(result[1].id, self.episode_2.id)
-
-    def test_episide_team_any_of_many(self):
-        """
-            test when looking for multiple tags only
-            episodes with all of those tags will be returned
-        """
-        self.episode_1.tagging_set.create(value="tree", archived=False)
-        self.episode_1.tagging_set.create(value="plant", archived=False)
-        self.episode_2.tagging_set.create(value="plant", archived=False)
-        self.episode_3.tagging_set.create(value="tree", archived=False)
-        query_end = dict(
-            query_type="Any Of",
-            value=["Plant", "Tree"],
-            field="team"
-        )
-
-        with patch.object(
-            search_rule_fields.models.Tagging, "build_field_schema"
-        ) as bfs:
-            bfs.return_value = [
-                dict(
-                    name="plant",
-                    title="Plant"
-                ),
-                dict(
-                    name="tree",
-                    title="Tree"
-                ),
-            ]
-            result = self.episode_rule.query(query_end)
-        self.assertEqual(len(result), 3)
-        self.assertEqual(result[0].id, self.episode_1.id)
-        self.assertEqual(result[1].id, self.episode_2.id)
-        self.assertEqual(result[2].id, self.episode_3.id)
-
-
 class LocationRuleTestCase(OpalTestCase):
     def test_location_rule_order(self):
         location_rule = search_rules.SearchRule.get_rule(
@@ -389,8 +202,12 @@ class EpisodeTestBase(OpalTestCase):
             fields[0].get_name(), "team"
         )
         self.assertEqual(
-            fields[1].get_name(), "start"
+            fields[1].get_name(), "category_name"
+        )
+
+        self.assertEqual(
+            fields[2].get_name(), "start"
         )
         self.assertEqual(
-            fields[2].get_name(), "end"
+            fields[3].get_name(), "end"
         )
